:recycle: (pc-15004) move EduconnectErros to adequate folder
