+++ src/features/navigation/RootNavigator/identityCheckRoutes.ts
@@ -6,10 +6,10 @@ import { NavigationSignUp } from 'features/cheatcodes/pages/NavigationSignUp'
 import { NavigationIdentityCheck } from 'features/cheatcodes/pages/NavigationSignUp/NavigationIdentityCheck'
 import { withAsyncErrorBoundary } from 'features/errors'
 import { withEduConnectErrorBoundary } from 'features/identityCheck/errors/eduConnect/EduConnectErrorBoundary'
+import { EduConnectErrors } from 'features/identityCheck/errors/eduConnect/EduConnectErrors'
 import { BeneficiaryRequestSent } from 'features/identityCheck/pages/confirmation/BeneficiaryRequestSent'
 import { IdentityCheckHonor } from 'features/identityCheck/pages/confirmation/IdentityCheckHonor'
 import { UnderageAccountCreated } from 'features/identityCheck/pages/confirmation/UnderageAccountCreated'
-import { EduConnectErrors } from 'features/identityCheck/pages/errors/EduConnectErrors'
 import { IdentityCheckDMS } from 'features/identityCheck/pages/identification/IdentityCheckDMS'
 import { IdentityCheckEduConnect } from 'features/identityCheck/pages/identification/IdentityCheckEduConnect'
 import { IdentityCheckEduConnectForm } from 'features/identityCheck/pages/identification/IdentityCheckEduConnectForm'
